refactor: rename config-manager UI component to config-settings-panel
